+++ tests/Compiler/ImportedAppsTest.php
@@ -9,7 +9,6 @@ use BEAR\AppMeta\Meta;
 use BEAR\Package\Exception\PharImportsUnreadableException;
 use BEAR\Package\Module\Import\ImportApp;
 use BEAR\Package\Module\ImportAppModule;
-use BEAR\Package\Module\WriteRule;
 use BEAR\Resource\Annotation\ImportAppConfig;
 use BEAR\Resource\Module\ResourceModule;
 use PHPUnit\Framework\TestCase;
@@ -92,7 +91,6 @@ class ImportedAppsTest extends TestCase
             protected function configure(): void
             {
                 $this->bind(AbstractAppMeta::class)->toInstance($this->meta);
-                $this->bind(WriteRule::class)->toInstance(new WriteRule($this->meta->name, 'app'));
             }
         };
     }